refactor: move vaults v1 into package
